04-deployment/02-streaming/azure_function/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
mlflow
azure-storage-blob
azure-identity